描画に関わるメソッドをcanvas.jsへ移行
